// frontend/src/api/generated/models/issueStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable issue tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */

export type IssueStatus = typeof IssueStatus[keyof typeof IssueStatus];


export const IssueStatus = {
  backlog: 'backlog',
  todo: 'todo',
  in_progress: 'in_progress',
  in_review: 'in_review',
  done: 'done',
  cancelled: 'cancelled',
} as const;